Fix Shadow Jar not including shadowed dependencies in the final build.
-Switch the shadowjar project configuration to use the shadow
configuration
-Disabled use of include dependency in shadowJar block. The change in
project configuration causes the include
dependency to break the shadow jar behaviour.
-Comment out dependency of jar process task in replaceResources task.
This fixes an issue that can occur where the task runs too late when
other projects depend on Techarium, resulting in the mods.toml file
being not populated.
-Added the --info flag to the github actions workflow build step. This
allows us to debug CI builds better in the future.
-Bumped Shadow Jar version to 6.1.0
-Update Immersive Engineering version to 5.0.2-137 to fix runtime crash
in a dev
environment caused by conflicting method name used by IE. This was
required in order to verify this shadow jar fix was working